Create WeatherAPI.js
I haven't converted this all to JSX yet, because I wasn't quite sure what our project's structure will look like/if this file will directly interact with the UI or just be called elsewhere, but here is the code for retrieving a forecast from the weather API, with a bunch of commented out other possible additions. Hopefully it makes sense.